pyproject.toml
[project]
name = "alphazero-chess"
version = "0.1.0"
description = "AlphaZero-style chess engine targeting 2200+ FIDE."
requires-python = ">=3.10"
dependencies = [
  "torch>=2.3",
  "python-chess>=1.999",
  "numpy>=1.26",
  "tqdm>=4.66",
  "zstandard>=0.22",
  "pyyaml>=6.0",
  "wandb>=0.17",
]

[project.optional-dependencies]
dev = ["pytest>=8", "ruff>=0.5"]

[tool.ruff]
line-length = 100
